pyproject.toml
[tool.poetry]
name = "acquisition-leadgen-engine"
version = "0.1.0"
description = "AI-powered lead generation and customer acquisition engine"
authors = ["AsifNewaz001"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
alembic = "^1.14.0"
httpx = "^0.27.2"
openai = "^1.55.0"
langchain = "^0.3.7"
python-dotenv = "^1.0.1"
celery = "^5.4.0"
redis = "^5.2.0"
hubspot-api-client = "^9.3.0"
simple-salesforce = "^1.12.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
flake8 = "^7.1.1"
mypy = "^1.13.0"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
asyncio_mode = "auto"
